+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Improvements
 
+* Fix integ test and integ test cluster #87
 * Node: Max volumes per node from organization quota #85
 * container-storage-interface: upgrade to v1.11.0 #73
 * egoscale: update to v3.1.9 to allow volumes of minimum size 1GiB #74