+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Improvements
+
+* Bump go.mk #113
+
 ## v0.31.2
 
 ### Improvements